Major cleanup and refactoring of verbs plugin
Treat expected an unexpected receives as essentially the same
and just maintain a list of receives posted without trying to
cope with which are which. Assign them when they are received.

Major simplification of whole receive handling code with only
wrinkle being that we need to check that clients always have
something preposted for when a message arrives.