Fix condition variable for OS threads and add debug of memory region delete
When HPX starts up, the parcelport init takes place on an OS thread, so we must
use a condition_variable_any and a spinlock rather than an hpx mutex.

When deleting a memory region (pushin back to the pool) we can add a debug mode
check that iterates over the pool to see if a double free/delet is taking place.